update variant submodule

+++ core/src/scene/sceneLoader.cpp
@@ -1135,7 +1135,7 @@ void SceneLoader::loadCamera(const Node& _camera, const std::shared_ptr<Scene>& 
             } else if (fov.IsSequence()) {
                 camera.fovStops = std::make_shared<Stops>(Stops::Numbers(fov));
                 for (auto& f : camera.fovStops->frames) {
-                    f.value = f.value.get<float>() * DEG_TO_RAD;
+                    f.value = float(f.value.get<float>() * DEG_TO_RAD);
                 }
             }
         }
